scrapper script updated
